bot/requirements.txt
aiogram==3.13.1
requests==2.32.3
python-dotenv==1.2.1
langchain-community
langchain-text-splitters
langchain-chroma
langchain_huggingface
sentence-transformers
torch
aiohttp